Struct unicode_script::ScriptExtension

#[non_exhaustive]pub struct ScriptExtension {
    first: u64,
    second: u64,
    third: u32,
    common: bool,
}

A value for the Script_Extension property

ScriptExtension is one or more Script

This is essentially an optimized version of Vec<Script> that uses bitfields

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
first: u64second: u64third: u32common: bool
Implementations

impl ScriptExtension

const THIRD_MAX: u32

pub(crate) const fn new(first: u64, second: u64, third: u32) -> Self

pub(crate) const fn new_common() -> Self

pub(crate) const fn new_inherited() -> Self

pub(crate) const fn new_unknown() -> Self

const fn is_common_or_inherited(self) -> bool

pub const fn is_common(self) -> bool

Checks if the script extension is Common

pub const fn is_inherited(self) -> bool

Checks if the script extension is Inherited

pub const fn is_empty(self) -> bool

Checks if the script extension is empty (unknown)

pub fn len(self) -> usize

Returns the number of scripts in the script extension

pub fn intersect_with(&mut self, other: Self)

Intersect this ScriptExtension with another ScriptExtension. Produces Unknown if things do not intersect. This is equivalent to ScriptExtension::intersection but it stores the result in self

"Common" (Zyyy) and "Inherited" (Zinh) are considered as intersecting everything, the intersection of Common and Inherited is Inherited

pub const fn intersection(self, other: Self) -> Self

Find the intersection between two ScriptExtensions. Returns Unknown if things do not intersect.

"Common" (Zyyy) and "Inherited" (Zinh) are considered as intersecting everything, the intersection of Common and Inherited is Inherited

pub const fn union(self, other: Self) -> Self

Find the union between two ScriptExtensions.

"Common" (Zyyy) and "Inherited" (Zinh) are considered as intersecting everything, the union of Common and Inherited is Common

pub fn contains_script(self, script: Script) -> bool

Check if this ScriptExtension contains the given script

Should be used with specific scripts only, this will return true if self is not Unknown and script is Common or Inherited

pub fn for_str(x: &str) -> Self

Get the intersection of script extensions of all characters in a string.

pub fn iter(self) -> ScriptIteratorⓘ

Important traits for ScriptIterator
impl Iterator for ScriptIterator type Item = Script;

Iterate over the scripts in this script extension

Will never yield Script::Unknown