Module unicode_script::tables::script_extensions

Constants

ADLAM
Adlam

ADLM
Adlam

ADLM_ARAB_MAND_MANI_PHLP_ROHG_SOGD_SYRC
Adlam, Arabic, Mandaic, Manichaean, Psalter_Pahlavi, Hanifi_Rohingya, Sogdian, Syriac

AGHB
Caucasian_Albanian

AHOM
Ahom

ANATOLIAN_HIEROGLYPHS
Anatolian_Hieroglyphs

ARAB
Arabic

ARABIC
Arabic

ARAB_COPT
Arabic, Coptic

ARAB_ROHG
Arabic, Hanifi_Rohingya

ARAB_ROHG_SYRC_THAA_YEZI
Arabic, Hanifi_Rohingya, Syriac, Thaana, Yezidi

ARAB_SYRC
Arabic, Syriac

ARAB_SYRC_THAA
Arabic, Syriac, Thaana

ARAB_THAA
Arabic, Thaana

ARAB_THAA_YEZI
Arabic, Thaana, Yezidi

ARMENIAN
Armenian

ARMI
Imperial_Aramaic

ARMN
Armenian

AVESTAN
Avestan

AVST
Avestan

BALI
Balinese

BALINESE
Balinese

BAMU
Bamum

BAMUM
Bamum

BASS
Bassa_Vah

BASSA_VAH
Bassa_Vah

BATAK
Batak

BATK
Batak

BENG
Bengali

BENGALI
Bengali

BENG_CAKM_SYLO
Bengali, Chakma, Syloti_Nagri

BENG_DEVA
Bengali, Devanagari

BENG_DEVA_DOGR_GONG_GONM_GRAN_GUJR_GURU_KNDA_LIMB_MAHJ_MLYM_NAND_ORYA_SIND_SINH_SYLO_TAKR_TAML_TELU_TIRH
Bengali, Devanagari, Dogra, Gunjala_Gondi, Masaram_Gondi, Grantha, Gujarati, Gurmukhi, Kannada, Limbu, Mahajani, Malayalam, Nandinagari, Oriya, Khudawadi, Sinhala, Syloti_Nagri, Takri, Tamil, Telugu, Tirhuta

BENG_DEVA_DOGR_GONG_GONM_GRAN_GUJR_GURU_KNDA_MAHJ_MLYM_NAND_ORYA_SIND_SINH_SYLO_TAKR_TAML_TELU_TIRH
Bengali, Devanagari, Dogra, Gunjala_Gondi, Masaram_Gondi, Grantha, Gujarati, Gurmukhi, Kannada, Mahajani, Malayalam, Nandinagari, Oriya, Khudawadi, Sinhala, Syloti_Nagri, Takri, Tamil, Telugu, Tirhuta

BENG_DEVA_GRAN_GUJR_GURU_KNDA_LATN_MLYM_ORYA_SHRD_TAML_TELU_TIRH
Bengali, Devanagari, Grantha, Gujarati, Gurmukhi, Kannada, Latin, Malayalam, Oriya, Sharada, Tamil, Telugu, Tirhuta

BENG_DEVA_GRAN_GUJR_GURU_KNDA_LATN_MLYM_ORYA_TAML_TELU_TIRH
Bengali, Devanagari, Grantha, Gujarati, Gurmukhi, Kannada, Latin, Malayalam, Oriya, Tamil, Telugu, Tirhuta

BENG_DEVA_GRAN_KNDA
Bengali, Devanagari, Grantha, Kannada

BENG_DEVA_GRAN_KNDA_NAND_ORYA_TELU_TIRH
Bengali, Devanagari, Grantha, Kannada, Nandinagari, Oriya, Telugu, Tirhuta

BHAIKSUKI
Bhaiksuki

BHKS
Bhaiksuki

BOPO
Bopomofo

BOPOMOFO
Bopomofo

BOPO_HANG_HANI_HIRA_KANA
Bopomofo, Hangul, Han, Hiragana, Katakana

BOPO_HANG_HANI_HIRA_KANA_YIII
Bopomofo, Hangul, Han, Hiragana, Katakana, Yi

BOPO_HANI
Bopomofo, Han

BRAH
Brahmi

BRAHMI
Brahmi

BRAI
Braille

BRAILLE
Braille

BUGI
Buginese

BUGINESE
Buginese

BUGI_JAVA
Buginese, Javanese

BUHD
Buhid

BUHD_HANO_TAGB_TGLG
Buhid, Hanunoo, Tagbanwa, Tagalog

BUHID
Buhid

CAKM
Chakma

CAKM_MYMR_TALE
Chakma, Myanmar, Tai_Le

CANADIAN_ABORIGINAL
Canadian_Aboriginal

CANS
Canadian_Aboriginal

CARI
Carian

CARIAN
Carian

CAUCASIAN_ALBANIAN
Caucasian_Albanian

CHAKMA
Chakma

CHAM
Cham

CHER
Cherokee

CHEROKEE
Cherokee

CHORASMIAN
Chorasmian

CHRS
Chorasmian

COMMON

COPT
Coptic

COPTIC
Coptic

CPRT
Cypriot

CPRT_LINA_LINB
Cypriot, Linear_A, Linear_B

CPRT_LINB
Cypriot, Linear_B

CUNEIFORM
Cuneiform

CYPRIOT
Cypriot

CYRILLIC
Cyrillic

CYRL
Cyrillic

CYRL_GLAG
Cyrillic, Glagolitic

CYRL_LATN
Cyrillic, Latin

CYRL_PERM
Cyrillic, Old_Permic

CYRL_SYRC
Cyrillic, Syriac

DESERET
Deseret

DEVA
Devanagari

DEVANAGARI
Devanagari

DEVA_DOGR_GUJR_GURU_KHOJ_KNDA_KTHI_MAHJ_MLYM_MODI_NAND_SIND_TAKR_TIRH
Devanagari, Dogra, Gujarati, Gurmukhi, Khojki, Kannada, Kaithi, Mahajani, Malayalam, Modi, Nandinagari, Khudawadi, Takri, Tirhuta

DEVA_DOGR_GUJR_GURU_KHOJ_KNDA_KTHI_MAHJ_MODI_NAND_SIND_TAKR_TIRH
Devanagari, Dogra, Gujarati, Gurmukhi, Khojki, Kannada, Kaithi, Mahajani, Modi, Nandinagari, Khudawadi, Takri, Tirhuta

DEVA_DOGR_GUJR_GURU_KHOJ_KTHI_MAHJ_MODI_SIND_TAKR_TIRH
Devanagari, Dogra, Gujarati, Gurmukhi, Khojki, Kaithi, Mahajani, Modi, Khudawadi, Takri, Tirhuta

DEVA_DOGR_KTHI_MAHJ
Devanagari, Dogra, Kaithi, Mahajani

DEVA_GRAN
Devanagari, Grantha

DEVA_GRAN_KNDA
Devanagari, Grantha, Kannada

DEVA_GRAN_LATN
Devanagari, Grantha, Latin

DEVA_KNDA_MLYM_ORYA_TAML_TELU
Devanagari, Kannada, Malayalam, Oriya, Tamil, Telugu

DEVA_NAND
Devanagari, Nandinagari

DEVA_SHRD
Devanagari, Sharada

DEVA_TAML
Devanagari, Tamil

DIAK
Dives_Akuru

DIVES_AKURU
Dives_Akuru

DOGR
Dogra

DOGRA
Dogra

DSRT
Deseret

DUPL
Duployan

DUPLOYAN
Duployan

EGYP
Egyptian_Hieroglyphs

EGYPTIAN_HIEROGLYPHS
Egyptian_Hieroglyphs

ELBA
Elbasan

ELBASAN
Elbasan

ELYM
Elymaic

ELYMAIC
Elymaic

ETHI
Ethiopic

ETHIOPIC
Ethiopic

GEOR
Georgian

GEORGIAN
Georgian

GEOR_LATN
Georgian, Latin

GLAG
Glagolitic

GLAGOLITIC
Glagolitic

GONG
Gunjala_Gondi

GONM
Masaram_Gondi

GOTH
Gothic

GOTHIC
Gothic

GRAN
Grantha

GRANTHA
Grantha

GRAN_TAML
Grantha, Tamil

GREEK
Greek

GREK
Greek

GUJARATI
Gujarati

GUJR
Gujarati

GUJR_KHOJ
Gujarati, Khojki

GUNJALA_GONDI
Gunjala_Gondi

GURMUKHI
Gurmukhi

GURU
Gurmukhi

GURU_MULT
Gurmukhi, Multani

HAN
Han

HANG
Hangul

HANGUL
Hangul

HANI
Han

HANIFI_ROHINGYA
Hanifi_Rohingya

HANI_HIRA_KANA
Han, Hiragana, Katakana

HANI_LATN
Han, Latin

HANO
Hanunoo

HANUNOO
Hanunoo

HATR
Hatran

HATRAN
Hatran

HEBR
Hebrew

HEBREW
Hebrew

HIRA
Hiragana

HIRAGANA
Hiragana

HIRA_KANA
Hiragana, Katakana

HLUW
Anatolian_Hieroglyphs

HMNG
Pahawh_Hmong

HMNP
Nyiakeng_Puachue_Hmong

HUNG
Old_Hungarian

IMPERIAL_ARAMAIC
Imperial_Aramaic

INHERITED

INSCRIPTIONAL_PAHLAVI
Inscriptional_Pahlavi

INSCRIPTIONAL_PARTHIAN
Inscriptional_Parthian

ITAL
Old_Italic

JAVA
Javanese

JAVANESE
Javanese

KAITHI
Kaithi

KALI
Kayah_Li

KALI_LATN_MYMR
Kayah_Li, Latin, Myanmar

KANA
Katakana

KANNADA
Kannada

KATAKANA
Katakana

KAYAH_LI
Kayah_Li

KHAR
Kharoshthi

KHAROSHTHI
Kharoshthi

KHITAN_SMALL_SCRIPT
Khitan_Small_Script

KHMER
Khmer

KHMR
Khmer

KHOJ
Khojki

KHOJKI
Khojki

KHUDAWADI
Khudawadi

KITS
Khitan_Small_Script

KNDA
Kannada

KNDA_NAND
Kannada, Nandinagari

KTHI
Kaithi

LANA
Tai_Tham

LAO
Lao

LAOO
Lao

LATIN
Latin

LATN
Latin

LATN_MONG
Latin, Mongolian

LEPC
Lepcha

LEPCHA
Lepcha

LIMB
Limbu

LIMBU
Limbu

LINA
Linear_A

LINB
Linear_B

LINEAR_A
Linear_A

LINEAR_B
Linear_B

LISU
Lisu

LYCI
Lycian

LYCIAN
Lycian

LYDI
Lydian

LYDIAN
Lydian

MAHAJANI
Mahajani

MAHJ
Mahajani

MAKA
Makasar

MAKASAR
Makasar

MALAYALAM
Malayalam

MAND
Mandaic

MANDAIC
Mandaic

MANI
Manichaean

MANICHAEAN
Manichaean

MARC
Marchen

MARCHEN
Marchen

MASARAM_GONDI
Masaram_Gondi

MEDEFAIDRIN
Medefaidrin

MEDF
Medefaidrin

MEETEI_MAYEK
Meetei_Mayek

MEND
Mende_Kikakui

MENDE_KIKAKUI
Mende_Kikakui

MERC
Meroitic_Cursive

MERO
Meroitic_Hieroglyphs

MEROITIC_CURSIVE
Meroitic_Cursive

MEROITIC_HIEROGLYPHS
Meroitic_Hieroglyphs

MIAO
Miao

MLYM
Malayalam

MODI
Modi

MONG
Mongolian

MONGOLIAN
Mongolian

MONG_PHAG
Mongolian, Phags_Pa

MRO
Mro

MROO
Mro

MTEI
Meetei_Mayek

MULT
Multani

MULTANI
Multani

MYANMAR
Myanmar

MYMR
Myanmar

NABATAEAN
Nabataean

NAND
Nandinagari

NANDINAGARI
Nandinagari

NARB
Old_North_Arabian

NBAT
Nabataean

NEWA
Newa

NEW_TAI_LUE
New_Tai_Lue

NKO
Nko

NKOO
Nko

NSHU
Nushu

NUSHU
Nushu

NYIAKENG_PUACHUE_HMONG
Nyiakeng_Puachue_Hmong

OGAM
Ogham

OGHAM
Ogham

OLCK
Ol_Chiki

OLD_HUNGARIAN
Old_Hungarian

OLD_ITALIC
Old_Italic

OLD_NORTH_ARABIAN
Old_North_Arabian

OLD_PERMIC
Old_Permic

OLD_PERSIAN
Old_Persian

OLD_SOGDIAN
Old_Sogdian

OLD_SOUTH_ARABIAN
Old_South_Arabian

OLD_TURKIC
Old_Turkic

OL_CHIKI
Ol_Chiki

ORIYA
Oriya

ORKH
Old_Turkic

ORYA
Oriya

OSAGE
Osage

OSGE
Osage

OSMA
Osmanya

OSMANYA
Osmanya

PAHAWH_HMONG
Pahawh_Hmong

PALM
Palmyrene

PALMYRENE
Palmyrene

PAUC
Pau_Cin_Hau

PAU_CIN_HAU
Pau_Cin_Hau

PERM
Old_Permic

PHAG
Phags_Pa

PHAGS_PA
Phags_Pa

PHLI
Inscriptional_Pahlavi

PHLP
Psalter_Pahlavi

PHNX
Phoenician

PHOENICIAN
Phoenician

PLRD
Miao

PRTI
Inscriptional_Parthian

PSALTER_PAHLAVI
Psalter_Pahlavi

REJANG
Rejang

RJNG
Rejang

ROHG
Hanifi_Rohingya

RUNIC
Runic

RUNR
Runic

SAMARITAN
Samaritan

SAMR
Samaritan

SARB
Old_South_Arabian

SAUR
Saurashtra

SAURASHTRA
Saurashtra

SGNW
SignWriting

SHARADA
Sharada

SHAVIAN
Shavian

SHAW
Shavian

SHRD
Sharada

SIDD
Siddham

SIDDHAM
Siddham

SIGNWRITING
SignWriting

SIND
Khudawadi

SINH
Sinhala

SINHALA
Sinhala

SOGD
Sogdian

SOGDIAN
Sogdian

SOGO
Old_Sogdian

SORA
Sora_Sompeng

SORA_SOMPENG
Sora_Sompeng

SOYO
Soyombo

SOYOMBO
Soyombo

SUND
Sundanese

SUNDANESE
Sundanese

SYLO
Syloti_Nagri

SYLOTI_NAGRI
Syloti_Nagri

SYRC
Syriac

SYRIAC
Syriac

TAGALOG
Tagalog

TAGB
Tagbanwa

TAGBANWA
Tagbanwa

TAI_LE
Tai_Le

TAI_THAM
Tai_Tham

TAI_VIET
Tai_Viet

TAKR
Takri

TAKRI
Takri

TALE
Tai_Le

TALU
New_Tai_Lue

TAMIL
Tamil

TAML
Tamil

TANG
Tangut

TANGUT
Tangut

TAVT
Tai_Viet

TELU
Telugu

TELUGU
Telugu

TFNG
Tifinagh

TGLG
Tagalog

THAA
Thaana

THAANA
Thaana

THAI
Thai

TIBETAN
Tibetan

TIBT
Tibetan

TIFINAGH
Tifinagh

TIRH
Tirhuta

TIRHUTA
Tirhuta

UGAR
Ugaritic

UGARITIC
Ugaritic

UNKNOWN

VAI
Vai

VAII
Vai

WANCHO
Wancho

WARA
Warang_Citi

WARANG_CITI
Warang_Citi

WCHO
Wancho

XPEO
Old_Persian

XSUX
Cuneiform

YEZI
Yezidi

YEZIDI
Yezidi

YI
Yi

YIII
Yi

ZANABAZAR_SQUARE
Zanabazar_Square

ZANB
Zanabazar_Square